BREAKING CHANGE: Added lazy loading option (closes #195)
There is a breaking change in `core/route.service.ts`:
`Route.withShell()` now returns a single route, so you have to update
your feature modules using it from:

```ts
const routes: Routes = Route.withShell([
  { path: 'about', component: AboutComponent, data: { title: extract('About') } }
]);
```

to:

```ts
const routes: Routes = [
  Route.withShell([
    { path: 'about', component: AboutComponent, data: { title: extract('About') } }
  ])
];
```
