--- generators/app/templates/src/app/about/about-routing.module.ts
@@ -1,16 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-
-import { Route, extract } from '@app/core';
-import { AboutComponent } from './about.component';
-
-const routes: Routes = Route.withShell([
-  { path: 'about', component: AboutComponent, data: { title: extract('About') } }
-]);
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-  providers: []
-})
-export class AboutRoutingModule { }
