+++ generators/app/templates/src/app/core/_route.service.ts
@@ -1,4 +1,4 @@
-import { Routes } from '@angular/router';
+import { Route as ngRoute, Routes } from '@angular/router';
 
 import { ShellComponent } from './shell/shell.component';
 <% if (props.auth) { -%>
@@ -13,10 +13,10 @@ export class Route {
   /**
    * Creates routes using the shell component and authentication.
    * @param routes The routes to add.
-   * @return {Routes} The new routes using shell as the base.
+   * @return {Route} The new route using shell as the base.
    */
-  static withShell(routes: Routes): Routes {
-    return [{
+  static withShell(routes: Routes): ngRoute {
+    return {
       path: '',
       component: ShellComponent,
       children: routes,
@@ -25,7 +25,7 @@ export class Route {
 <% } -%>
       // Reuse ShellComponent instance when navigating between child views
       data: { reuse: true }
-}];
+    };
   }
 
 }
